move framework functions to node-utils
