+++ packages/core/__tests__/node-utils/framework.test.ts
@@ -1,10 +1,10 @@
 import { describe, expect, test, jest, afterEach } from "@jest/globals";
-import { Frameworks, SupportedFrameworks } from "../../../src/models/Framework";
-import { JsonPackage } from "../../../src/filesystem/pkg";
-import * as FrameworkUtils from "../../../src/utils/framework";
+import { Frameworks, SupportedFrameworks } from "../../src/models/Framework";
+import { JsonPackage } from "../../src/filesystem/pkg";
+import * as FrameworkUtils from "../../src/node-utils/framework";
 import * as fs from "fs";
 import { mocked } from "ts-jest/utils";
-import { Manifest } from "../../../src/models/Manifest";
+import { Manifest } from "../../src/models/Manifest";
 
 jest.mock("fs");
