+++ packages/core/src/index.ts
@@ -6,3 +6,4 @@ export * as Libraries from "./libraries";
 export * as Mocks from "./libraries/mocks";
 export * from "./auth";
 export * as CONSTS from "./consts";
+export * as NodeUtils from "./node-utils";
