+++ packages/init/__tests__/configure-project.test.ts
@@ -68,7 +68,7 @@ describe("configure-project", () => {
   const fakeBase = "https://music.to.my.hears.io" as Core.Utils.Endpoints.Base;
   const fakeRepository = "testing-repo";
   const fakeFrameworkStats = {
-    value: Core.Utils.Framework.Frameworks.react,
+    value: Core.Models.Frameworks.react,
     manuallyAdded: false,
   };
 
@@ -107,7 +107,7 @@ describe("configure-project", () => {
     retrieveManifestMock.mockReturnValue({
       exists: true,
       content: {
-        framework: Core.Utils.Framework.Frameworks.react,
+        framework: Core.Models.Frameworks.react,
       },
     });
     addJsonPackageSmScriptMock.mockReturnValue(true);
@@ -130,7 +130,7 @@ describe("configure-project", () => {
     retrieveManifestMock.mockReturnValue({
       exists: true,
       content: {
-        framework: Core.Utils.Framework.Frameworks.react,
+        framework: Core.Models.Frameworks.react,
       },
     });
     addJsonPackageSmScriptMock.mockReturnValue(true);
@@ -190,7 +190,7 @@ describe("configure-project", () => {
     retrieveManifestMock.mockReturnValue({
       exists: true,
       content: {
-        framework: Core.Utils.Framework.Frameworks.react,
+        framework: Core.Models.Frameworks.react,
       },
     });
     patchManifestMock.mockImplementation(() => {
